Update sentora_php7_pkg_update.sh
Fixed php.ini rewrite loop for CentOS

+++ sentora_php7_pkg_update.sh
@@ -378,8 +378,11 @@ done
 	if [[ "$OS" = "CentOs" && ( "$VER" = "6" | "$VER" = "7" ) ]]; then
 	
 		## Setup PHP 7.3 new PHP.INI file shipped with PHP and rename old PHP.INI
-		mv /etc/php.ini /etc/php.ini.OLD
-		cp -r /etc/php.ini.rpmnew /etc/php.ini
+		file="/etc/php.ini.OLD"
+		if [ ! -f "$file" ]; then
+			mv /etc/php.ini /etc/php.ini.OLD
+			cp -r /etc/php.ini.rpmnew /etc/php.ini
+		fi
 	
 		# Fix missing php.ini settings sentora needs
 		echo -e "\nFix missing php.ini settings sentora needs in CentOS 6.x php 7.3 ..."
